Panelise Trilat
Trilat to panel code..
Distance parser to base utils.
Add in pop outable trilat panel
#region trilat code, removed the special form and used message box
instead
fixed some issues with wrong namespacing in usercontrols..
withdrawn an old edd.db namespace
