+++ EDDiscovery/Properties/Resources.resx
@@ -385,4 +385,7 @@
   <data name="expedition" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\expedition.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
+  <data name="triangulation" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\triangulation.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
 </root>
