+++ EDDiscovery/UserControls/UserControlRouteTracker.cs
@@ -23,7 +23,6 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 using EliteDangerousCore.DB;
-using EDDiscovery.DB;
 using EliteDangerousCore;
 
 namespace EDDiscovery.UserControls
